src/add-practice/requirements.txt
google-cloud-aiplatform
google-cloud-storage
python-dotenv
jsonschema